requirements.txt
streamlit==1.54.0
pypdf==5.3.0
python-docx==1.2.0
python-dotenv==1.0.1
openai==1.99.9
